Fix io expander label and add udev-toradex-rules.bbappend (#51)
On the Rev B RCU, the mem address of I2C is different from that of Rev A. Hence, the label on the IO Expander should reflect the new mem address.
Because we use the original apalis-imx8 MACHINE instead of creating a custom "smartracks-imx8" MACHINE, bitbake bakes the udev-toradex-rules recipe. In Rev A, the 99-toradex.rules is an empty file because it takes from https://git.toradex.com/cgit/meta-toradex-bsp-common.git/tree/recipes-core/udev/files/99-toradex.rules?h=dunfell-5.x.y instead from the folder within the MACHINE. The 99-toradex.rules from apalis-imx8 defines a set of apalis specific SYMLINKs which are not applicable to smartracks (and makes the rules untidy). Hence, a bbappend is added to remove 99-toradex.rules entirely.

* add line to remove 99-toradex.rules and edit IO expander name

Signed-off-by: wkoe <wilson.koe@ni.com>

* move removing file to do_install_append

* change it to udev-extraconf

* add udev-toradex-rules.bbappend

* address feedbacks by adding comments and some udev rules